Updated Foodr to Version 0.6
Changes: Added the ability to delete your favourites. By extension this
cascade deletes any review a user has made about a location, also if
that favourite is the only user to favourite a certain place, then the
place information is deleted from the database too.

Added the ability to review as well as view reviews made by yourself and
other people in the place detail page.

Known bugs:
The user has to wait for a few seconds during start up before doing any
input, otherwise the app will not respond.

The open now in the place detail is currently broken and will say a
place is always open.